fix(arch): resolve remaining layer contract violations
Move FeatureFlagService from src/app/services to src/infra/services
since it owns UoW + ORM model — this is infra-level code.

Move enqueue_pending_images UoW logic into a new
src/infra/services/pending_image_enqueue_service so that
api/dependencies/meal_image_cache no longer imports uow_async.

Re-export BackgroundTaskManager from src/infra/event_bus __init__
so api/dependencies/task_manager imports from the package, not the
submodule — satisfying the checker's allowlist.

Contracts: 4 kept, 0 broken.
