+++ src/api/dependencies/meal_image_cache.py
@@ -7,27 +7,28 @@ for image-text scoring, which requires vision encoding.
 
 from __future__ import annotations
 
-import logging
-
 from fastapi import Depends
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from src.domain.model.meal_image_cache import PendingItem
 from src.domain.services.meal_image_cache.meal_image_cache_service import (
     MealImageCacheService,
 )
 from src.infra.adapters.gemini_text_embedding_adapter import get_gemini_text_embedder
 from src.infra.config.settings import get_settings
 from src.infra.database.config_async import get_async_db
-from src.infra.database.uow_async import AsyncUnitOfWork
 from src.infra.repositories.pending_meal_image_repository_async import (
     AsyncPendingMealImageRepository,
 )
 from src.infra.repositories.pgvector_meal_image_cache_repository_async import (
     AsyncPgvectorMealImageCacheRepository,
 )
+from src.infra.services.pending_image_enqueue_service import enqueue_pending_images
 
-logger = logging.getLogger(__name__)
+__all__ = [
+    "get_meal_image_cache_service",
+    "get_pending_queue",
+    "enqueue_pending_images",
+]
 
 
 async def get_meal_image_cache_service(
@@ -47,12 +48,3 @@ async def get_pending_queue(
     return AsyncPendingMealImageRepository(session)
 
 
-async def enqueue_pending_images(items: list[PendingItem]) -> None:
-    """Enqueue pending image items in their own UoW transaction (non-fatal)."""
-    if not items:
-        return
-    try:
-        async with AsyncUnitOfWork() as uow:
-            await AsyncPendingMealImageRepository(uow.session).enqueue_many(items)
-    except Exception as e:
-        logger.warning("Pending image queue enqueue failed (non-fatal): %s", e)
