+++ src/infra/event_bus/__init__.py
@@ -2,7 +2,8 @@
 Event bus infrastructure implementation.
 """
 
+from .background_task_manager import BackgroundTaskManager
 from .event_bus import EventBus
 from .pymediator_event_bus import PyMediatorEventBus
 
-__all__ = ["EventBus", "PyMediatorEventBus"]
+__all__ = ["BackgroundTaskManager", "EventBus", "PyMediatorEventBus"]
